feat: use BHP256 as default hash function
tests: reflect hash changes

clippy: misc lints

ref: split crypto lib into mods

ref: remove unused imports

ref: move `CurrentNetwork` back into `lib.rs`
