+++ narwhal/consensus/src/tests/consensus_tests.rs
@@ -3,7 +3,7 @@
 
 #![allow(clippy::mutable_key_type)]
 
-use fastcrypto::hash::Hash;
+use crypto::Hash;
 use std::collections::BTreeSet;
 use storage::NodeStorage;
 use telemetry_subscribers::TelemetryGuards;
