+++ narwhal/primary/src/block_remover.rs
@@ -4,8 +4,7 @@ use crate::utils;
 use anyhow::Result;
 use config::{SharedWorkerCache, WorkerId};
 use consensus::dag::{Dag, ValidatorDagError};
-use crypto::PublicKey;
-use fastcrypto::hash::Hash;
+use crypto::{Hash, PublicKey};
 use futures::future::try_join_all;
 use itertools::Either;
 use network::PrimaryToWorkerRpc;
