+++ narwhal/primary/src/block_synchronizer/mock.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 use crate::block_synchronizer::{BlockHeader, BlockSynchronizeResult, Command};
-use fastcrypto::hash::Hash;
+use crypto::Hash;
 use std::collections::HashMap;
 use tokio::sync::{mpsc, oneshot};
 use types::CertificateDigest;
