+++ narwhal/primary/src/block_synchronizer/tests/handler_tests.rs
@@ -8,7 +8,7 @@ use crate::{
     common::create_db_stores,
     BlockHeader, MockBlockSynchronizer,
 };
-use fastcrypto::hash::Hash;
+use crypto::Hash;
 use std::{collections::HashSet, time::Duration};
 use test_utils::{fixture_payload, CommitteeFixture};
 use types::CertificateDigest;
