+++ narwhal/primary/src/tests/block_remover_tests.rs
@@ -4,8 +4,8 @@ use crate::{block_remover::BlockRemover, common::create_db_stores, NUM_SHUTDOWN_
 use anemo::PeerId;
 use config::{Committee, WorkerId};
 use consensus::dag::Dag;
-use crypto::{traits::KeyPair, PrivateKey};
-use fastcrypto::hash::Hash;
+use crypto::{Hash, PrivateKey};
+use fastcrypto::traits::KeyPair;
 use futures::future::join_all;
 use std::{borrow::Borrow, collections::HashMap, sync::Arc, time::Duration};
 use test_utils::CommitteeFixture;
