+++ narwhal/primary/src/tests/block_waiter_tests.rs
@@ -6,8 +6,8 @@ use crate::{
     BlockWaiter,
 };
 use anemo::PeerId;
-use crypto::traits::KeyPair as _;
-use fastcrypto::hash::Hash;
+use crypto::Hash;
+use fastcrypto::traits::KeyPair;
 use mockall::*;
 use std::sync::Arc;
 use test_utils::{
@@ -268,7 +268,7 @@ async fn test_return_error_when_certificate_is_missing() {
     let name = primary.public_key();
 
     // AND create a certificate but don't store it
-    let digest = CertificateDigest::new([0u8; 32]);
+    let digest = CertificateDigest::default();
 
     // AND mock the responses of the BlockSynchronizer
     let mut mock_handler = MockHandler::new();
@@ -308,7 +308,7 @@ async fn test_return_error_when_certificate_is_missing_when_get_blocks() {
     let name = primary.public_key();
 
     // AND create a certificate but don't store it
-    let digest = CertificateDigest::new([0u8; 32]);
+    let digest = CertificateDigest::default();
 
     // AND mock the responses of the BlockSynchronizer
     let mut mock_handler = MockHandler::new();
