+++ narwhal/primary/tests/integration_tests_proposer_api.rs
@@ -5,8 +5,8 @@ use arc_swap::ArcSwap;
 use bytes::Bytes;
 use config::{Epoch, Parameters};
 use consensus::dag::Dag;
-use crypto::PublicKey;
-use fastcrypto::{hash::Hash, traits::KeyPair as _};
+use crypto::{Hash, PublicKey};
+use fastcrypto::traits::KeyPair as _;
 use narwhal_primary as primary;
 use narwhal_primary::NUM_SHUTDOWN_RECEIVERS;
 use primary::{NetworkModel, Primary, CHANNEL_CAPACITY};
