+++ narwhal/types/src/error.rs
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 use crate::{HeaderDigest, Round, TimestampMs, VoteDigest};
 use config::Epoch;
-use fastcrypto::hash::Digest;
+use crypto::Digest;
 use store::StoreError;
 use thiserror::Error;
 
@@ -80,13 +80,13 @@ pub enum DagError {
     HeaderRequiresQuorum(HeaderDigest),
 
     #[error("Message {0} (round {1}) too old for GC round {2}")]
-    TooOld(Digest<{ crypto::DIGEST_LENGTH }>, Round, Round),
+    TooOld(Digest, Round, Round),
 
     #[error("Message {0} (round {1}) is too new for this primary at round {2}")]
-    TooNew(Digest<{ crypto::DIGEST_LENGTH }>, Round, Round),
+    TooNew(Digest, Round, Round),
 
     #[error("Vote {0} (round {1}) too old for round {2}")]
-    VoteTooOld(Digest<{ crypto::DIGEST_LENGTH }>, Round, Round),
+    VoteTooOld(Digest, Round, Round),
 
     #[error("Invalid epoch (expected {expected}, received {received})")]
     InvalidEpoch { expected: Epoch, received: Epoch },
