+++ narwhal/worker/src/batch_maker.rs
@@ -1,7 +1,7 @@
 // Copyright (c) 2021, Facebook, Inc. and its affiliates
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
-use fastcrypto::hash::Hash;
+use crypto::Hash;
 use futures::stream::FuturesOrdered;
 use store::Store;
